requirements.txt
metatrader5
pandas
numpy
python-dotenv
python-telegram-bot
scipy